flatten: trailing_text
