+++ packages/client/lib/app/core/widgets/trailing_text/mobx/trailing_text_store.dart
@@ -4,7 +4,7 @@ import 'package:mobx/mobx.dart';
 import 'package:nokhte/app/core/interfaces/logic.dart';
 import 'package:nokhte/app/core/mobx/mobx.dart';
 import 'package:nokhte/app/core/types/types.dart';
-import 'package:nokhte/app/core/widgets/trailing_text/stack/movies/reveal_text_movie.dart';
+import 'package:nokhte/app/core/widgets/trailing_text/movies/reveal_text_movie.dart';
 import 'package:nokhte/app/core/widgets/widgets.dart';
 import 'package:simple_animations/simple_animations.dart';
 part 'trailing_text_store.g.dart';
