+++ packages/client/lib/app/core/widgets/trailing_text/movies/reveal_text_movie.dart
