--- packages/client/lib/app/core/widgets/trailing_text/stack/constants/constants.dart
@@ -1,2 +1,0 @@
-export 'data/data.dart';
-export 'types/types.dart';
