--- packages/client/lib/app/core/widgets/trailing_text/stack/presentation/presentation.dart
@@ -1,2 +1,0 @@
-export 'mobx/trailing_text_store.dart';
-export 'widget/trailing_text.dart';
