+++ packages/client/lib/app/core/widgets/trailing_text/types/text_reveal_direction.dart
