+++ packages/client/lib/app/core/widgets/trailing_text/types/types.dart
