+++ packages/client/lib/app/core/widgets/trailing_text/widget/hidden_side.dart
